+++ CHANGELOG.md
@@ -6,8 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Changed
-- Upgrade OAuth endpoints
-- Update available scopes
+- Upgrade OAuth flow and parse JWT data
+- Update list of available scopes
 
 ## [v1.1.0] - 2024-10-21
 ### Changed