+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [v2.0.0] - 2024-10-22
 ### Changed
 - Upgrade OAuth flow and parse JWT data
 - Update list of available scopes
@@ -33,7 +35,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Everything. Initial alpha release
 
-[Unreleased]: https://github.com/timothydc/laravel-lightspeed-retail-api/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/timothydc/laravel-lightspeed-retail-api/compare/v2.0.0...HEAD
+[v2.0.0]: https://github.com/timothydc/laravel-lightspeed-retail-api/compare/v1.1.0...v2.0.0
 [v1.1.0]: https://github.com/timothydc/laravel-lightspeed-retail-api/compare/v1.0.0...v1.1.0
 [v1.0.0]: https://github.com/timothydc/laravel-lightspeed-retail-api/compare/v0.21-alpha...v1.0.0
 [v0.21-alpha]: https://github.com/timothydc/laravel-lightspeed-retail-api/compare/v0.20-alpha...v0.21-alpha